backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
aif360==0.6.1
google-generativeai==0.5.4
reportlab==4.2.0
python-dotenv==1.0.1
scipy==1.13.0
pydantic==2.7.1
httpx==0.27.0
